Split tests for helpers into separate modules
